backend_parser/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic>=2.0.0
pydantic-settings
PyJWT==2.8.0
websockets>=11.0.3